CLN: move pandas/tseries/resample.py -> pandas/core/resample.py
xref #13634
