+++ pandas/core/api.py
@@ -30,6 +30,7 @@ from pandas.tseries.index import (DatetimeIndex, Timestamp,
                                   date_range, bdate_range)
 from pandas.tseries.tdi import TimedeltaIndex, Timedelta
 from pandas.tseries.period import Period, PeriodIndex
+from pandas.core.resample import TimeGrouper
 
 # see gh-14094.
 from pandas.util.depr_module import _DeprecatedModule
