+++ pandas/core/resample.py
