+++ pandas/tests/test_resample.py
@@ -22,8 +22,8 @@ from pandas.tseries.frequencies import to_offset
 from pandas.tseries.index import date_range
 from pandas.tseries.offsets import Minute, BDay
 from pandas.tseries.period import period_range, PeriodIndex, Period
-from pandas.tseries.resample import (DatetimeIndex, TimeGrouper,
-                                     DatetimeIndexResampler)
+from pandas.core.resample import (DatetimeIndex, TimeGrouper,
+                                  DatetimeIndexResampler)
 from pandas.tseries.tdi import timedelta_range, TimedeltaIndex
 from pandas.util.testing import (assert_series_equal, assert_almost_equal,
                                  assert_frame_equal, assert_index_equal)
