+++ pandas/tseries/api.py
@@ -8,7 +8,6 @@ from pandas.tseries.index import DatetimeIndex, date_range, bdate_range
 from pandas.tseries.frequencies import infer_freq
 from pandas.tseries.tdi import Timedelta, TimedeltaIndex, timedelta_range
 from pandas.tseries.period import Period, PeriodIndex, period_range, pnow
-from pandas.tseries.resample import TimeGrouper
 from pandas.tseries.timedeltas import to_timedelta
 from pandas._libs.lib import NaT
 import pandas.tseries.offsets as offsets
